feat: create namespace for token permissions
create `tokenPermissions` namespace in the `Identity` Entity to handle all External Agents
  functionality

BREAKING CHANGE:
- move `identity.hasTokenPermissions` to `identity.tokenPermissions.hasPermissions`
- move `identity.getTokenPermissions` to `identity.tokenPermissions.get`
- move `identity.waivePermissions` to `identity.tokenPermissions.waive`
- change the type of `AgentWithGroup.agent` from `Agent` to `Identity`. This affects the return
  type of `securityToken.permissions.getAgents`
- remove the `Agent` entity:
    - move `agent.setPermissionGroup` to `identity.tokenPermissions.setGroup`. It now also
      receives the token as a parameter when not passing an existing group. It now throws
      an error if the target is not already an agent
    - move `agent.getPermissionGroup` to `identity.tokenPermissions.getGroup`. It now also
      receives the token as a parameter
    - move `agent.addedAt` to `identity.tokenPermissions.grantedAt`. It now also receives
      the token as a parameter
