+++ src/api/procedures/setPermissionGroup.ts
@@ -2,7 +2,6 @@ import { AgentGroup, TxTags } from 'polymesh-types/types';
 
 import { isFullGroupType } from '~/api/procedures/utils';
 import {
-  Agent,
   Context,
   createGroup,
   CustomPermissionGroup,
@@ -12,35 +11,44 @@ import {
   Procedure,
   SecurityToken,
 } from '~/internal';
-import { ErrorCode, TransactionPermissions, TxGroup } from '~/types';
+import { ErrorCode, Identity, isEntity, TransactionPermissions, TxGroup } from '~/types';
 import { MaybePostTransactionValue, ProcedureAuthorization } from '~/types/internal';
 import {
   permissionGroupIdentifierToAgentGroup,
   stringToIdentityId,
   stringToTicker,
 } from '~/utils/conversion';
+import { getToken } from '~/utils/internal';
 
 export interface SetPermissionGroupParams {
   /**
-   * Permission Group to assign to the Agent. Optionally, transaction permissions can be passed directly.
-   *   In that case, a new Permission Group will be created and consequently assigned to the Agent
+   * Permission Group to assign to the Identity. The Security Token is inferred from the group.
+   *   Optionally, transaction permissions and a Security Token can be passed directly.
+   *   In that case, a new Permission Group will be created and consequently assigned to the Identity
    */
   group:
     | KnownPermissionGroup
     | CustomPermissionGroup
-    | {
-        transactions: TransactionPermissions;
-      }
-    | {
-        transactionGroups: TxGroup[];
-      };
+    | ({
+        /**
+         * Security Token over which the Identity will be granted permissions
+         */
+        token: string | SecurityToken;
+      } & (
+        | {
+            transactions: TransactionPermissions;
+          }
+        | {
+            transactionGroups: TxGroup[];
+          }
+      ));
 }
 
 /**
  * @hidden
  */
 export type Params = SetPermissionGroupParams & {
-  agent: Agent;
+  identity: Identity;
 };
 
 /**
@@ -79,11 +87,12 @@ export async function prepareSetPermissionGroup(
     storage: { token },
   } = this;
 
-  const { agent, group } = args;
-  const { ticker, did } = agent;
+  const { identity, group } = args;
+  const { ticker } = token;
+  const { did } = identity;
 
   const [currentGroup, currentAgents] = await Promise.all([
-    agent.getPermissionGroup(),
+    identity.tokenPermissions.getGroup({ token }),
     token.permissions.getAgents(),
   ]);
 
@@ -100,10 +109,17 @@ export async function prepareSetPermissionGroup(
     }
   }
 
+  if (!currentAgents.find(({ agent }) => agent.did === did)) {
+    throw new PolymeshError({
+      code: ErrorCode.ValidationError,
+      message: 'The target must already be an Agent for the Security Token',
+    });
+  }
+
   let returnValue: MaybePostTransactionValue<CustomPermissionGroup | KnownPermissionGroup>;
   let rawAgentGroup;
 
-  if ('transactions' in group || 'transactionGroups' in group) {
+  if (!isEntity(group)) {
     returnValue = await this.addProcedure(createGroup(), {
       ticker,
       permissions: group,
@@ -154,12 +170,20 @@ export function getAuthorization(
  */
 export function prepareStorage(
   this: Procedure<Params, CustomPermissionGroup | KnownPermissionGroup, Storage>,
-  { agent: { ticker } }: Params
+  { group }: Params
 ): Storage {
   const { context } = this;
 
+  let token: string | SecurityToken;
+
+  if (isEntity(group)) {
+    ({ ticker: token } = group);
+  } else {
+    ({ token } = group);
+  }
+
   return {
-    token: new SecurityToken({ ticker }, context),
+    token: getToken(token, context),
   };
 }
