+++ src/internal.ts
@@ -126,7 +126,6 @@ export { NumberedPortfolio } from '~/api/entities/NumberedPortfolio';
 export { TransactionQueue } from '~/base/TransactionQueue';
 export { Checkpoint } from '~/api/entities/Checkpoint';
 export { CheckpointSchedule } from '~/api/entities/CheckpointSchedule';
-export { Agent } from '~/api/entities/Agent';
 export { PermissionGroup } from '~/api/entities/PermissionGroup';
 export { KnownPermissionGroup } from '~/api/entities/KnownPermissionGroup';
 export { CustomPermissionGroup } from '~/api/entities/CustomPermissionGroup';
